Added /W4 and fixed most the warnings

+++ src/gl_renderer.cpp
@@ -476,7 +476,7 @@ internal void renderer_resize()
 {
   if(glContext.initialized)
   {
-    glViewport(0, 0, input->screenSize.x, input->screenSize.y);
+    glViewport(0, 0, (int)input->screenSize.x, (int)input->screenSize.y);
     glUniform2fv(glContext.screenSizeID, 1, (float*)&input->screenSize);
   }
 }
